refactor(recoil): 디렉터리 구조 변경
orderConfirm, shoppingCart 디렉터리에서 각각의 도메인에 맞게 atom, selector 이동
