+++ src/recoil/orderConfirm/atoms.ts
@@ -1,24 +1,4 @@
-import OrderConfirmFetcher from '@apis/orderConfirm';
-import { Coupon } from '@appTypes/orderConfirm';
-import { atom, selector } from 'recoil';
-
-export const couponListSelector = selector<Coupon[]>({
-  key: 'couponListSelector',
-  get: async () => {
-    const coupons = await OrderConfirmFetcher.getCoupons();
-    return coupons;
-  },
-});
-
-export const couponListAtom = atom<Coupon[]>({
-  key: 'couponListAtom',
-  default: couponListSelector,
-});
-
-export const selectedCouponListAtom = atom<Coupon[]>({
-  key: 'selectedCouponListAtom',
-  default: [],
-});
+import { atom } from 'recoil';
 
 export const isInaccessibleAreaAtom = atom({
   key: 'isInaccessibleAreaAtom',
