+++ src/recoil/orderConfirm/discount.ts
@@ -1,6 +1,7 @@
 import { calculateDiscountAmount } from '@domain/discount';
-import { selectedCouponListAtom } from '@recoil/orderConfirm/atoms';
-import { orderPriceSelector, selectedItemsSelector, shippingPriceSelector } from '@recoil/shoppingCart';
+import { selectedCouponListAtom } from '@recoil/orderConfirm/coupon';
+import { orderPriceSelector, shippingPriceSelector } from '@recoil/shoppingCart';
+import { selectedItemsSelector } from '@recoil/shoppingCart/cartItems';
 import { selector } from 'recoil';
 
 export const totalDiscountPriceSelector = selector({
@@ -11,14 +12,10 @@ export const totalDiscountPriceSelector = selector({
     const selectedCartItems = get(selectedItemsSelector);
     const orderPrice = get(orderPriceSelector);
 
-    return selectedCouponList.reduce((acc, coupon) => {
-      const discountAmount = calculateDiscountAmount({
-        coupon,
-        selectedCartItems,
-        shippingPrice,
-        orderPrice,
-      });
-      return acc + discountAmount;
-    }, 0);
+    return selectedCouponList.reduce(
+      (prevDiscountPrice, coupon) =>
+        calculateDiscountAmount({ coupon, shippingPrice, orderPrice, selectedCartItems }) + prevDiscountPrice,
+      0,
+    );
   },
 });
