+++ src/recoil/shoppingCart/index.ts
@@ -1,2 +1,3 @@
-export * from './atoms';
-export * from './selectors';
+export * from './price';
+export * from './cartItems';
+export * from './quantity';
